+++ lib/cart/cart_screen.dart
@@ -1,0 +1,53 @@
+import 'package:cart_sample/cart/widget/add_more_widget.dart';
+import 'package:cart_sample/cart/widget/billing_widget.dart';
+import 'package:cart_sample/cart/widget/menu_widget.dart';
+import 'package:cart_sample/cart/widget/order_button_widget.dart';
+import 'package:cart_sample/cart/widget/store_name_widget.dart';
+import 'package:flutter/material.dart';
+
+class CartScreen extends StatefulWidget {
+  const CartScreen({Key? key}) : super(key: key);
+
+  @override
+  State<CartScreen> createState() => _CartScreenState();
+}
+
+class _CartScreenState extends State<CartScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: const Color.fromRGBO(246, 246, 246, 1.0),
+      appBar: AppBar(
+        leading: const BackButton(
+          color: Colors.black,
+        ),
+        title: const Text(
+          '장바구니',
+          style: TextStyle(
+            color: Colors.black,
+          ),
+        ),
+        elevation: 0,
+        backgroundColor: Colors.white,
+      ),
+      body: ListView(
+        children: [
+          SizedBox(
+            height: 10,
+          ),
+          StoreNameWidget(),
+          SizedBox(
+            height: 1,
+          ),
+          MenuWidget(),
+          SizedBox(
+            height: 1,
+          ),
+          AddMoreWidget(),
+          BillingWidget(),
+        ],
+      ),
+      bottomNavigationBar: OrderButtonWidget(),
+    );
+  }
+}
