+++ lib/cart/widget/order_button_widget.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+
+class OrderButtonWidget extends StatelessWidget {
+  const OrderButtonWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.white,
+      child: SafeArea(
+        child: Container(
+          height: 65,
+          padding: const EdgeInsets.symmetric(
+            horizontal: 20,
+            vertical: 10,
+          ),
+          child: ElevatedButton(
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Container(
+                  width: 25,
+                  height: 25,
+                  decoration: BoxDecoration(
+                    color: Colors.white,
+                    shape: BoxShape.circle,
+                  ),
+                  child: Center(
+                    child: Text(
+                      '1',
+                      style: TextStyle(
+                        color: Color.fromRGBO(44, 191, 188, 1.0),
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ),
+                SizedBox(
+                  width: 7,
+                ),
+                Text(
+                  '21,000원 배달 주문하기',
+                  style: TextStyle(
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ],
+            ),
+            style: ButtonStyle(
+              backgroundColor: MaterialStateProperty.all(
+                Color.fromRGBO(44, 191, 188, 1.0),
+              ),
+            ),
+            onPressed: () {},
+          ),
+        ),
+      ),
+    );
+  }
+}
